refactor(web): keep the folder icon with the row it belongs to
A file of its own was more than one hook needed. It lives with the row that
mainly uses it, and the popover imports it from there.

It stays out of the barrel either way — the two rows share it, but nothing
outside components/ has any use for it.
